refactor(diff): move MySQL repair into plugin
