+++ chat2db-community-server/chat2db-community-plugins/chat2db-community-mysql/pom.xml
@@ -18,6 +18,10 @@
             <artifactId>chat2db-community-spi</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.liquibase</groupId>
+            <artifactId>liquibase-core</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter</artifactId>
             <scope>test</scope>
