+++ chat2db-community-server/chat2db-community-plugins/chat2db-community-mysql/src/main/java/ai/chat2db/plugin/mysql/MysqlPlugin.java
@@ -1,7 +1,9 @@
 package ai.chat2db.plugin.mysql;
 
 import ai.chat2db.plugin.mysql.account.MysqlAccountManager;
+import ai.chat2db.plugin.mysql.diff.MysqlDiffChangeSetProcessor;
 import ai.chat2db.spi.IAccountManager;
+import ai.chat2db.spi.IDbDiffChangeSetProcessor;
 import ai.chat2db.spi.IDbManager;
 import ai.chat2db.spi.IDbMetaData;
 import ai.chat2db.spi.IPlugin;
@@ -10,6 +12,9 @@ import ai.chat2db.community.domain.api.config.DBConfig;
 import ai.chat2db.spi.util.FileUtils;
 
 public class MysqlPlugin extends MysqlSyntaxPlugin implements IPlugin {
+
+    private static final IDbDiffChangeSetProcessor DIFF_CHANGE_SET_PROCESSOR =
+            new MysqlDiffChangeSetProcessor();
 
     private DBConfig dbConfig;
 
@@ -30,6 +35,11 @@ public class MysqlPlugin extends MysqlSyntaxPlugin implements IPlugin {
     @Override
     public IDbManager getDbManager() {
         return new MysqlDBManager();
+    }
+
+    @Override
+    public IDbDiffChangeSetProcessor getDiffChangeSetProcessor() {
+        return DIFF_CHANGE_SET_PROCESSOR;
     }
 
     @Override
