+++ chat2db-community-server/chat2db-community-spi/pom.xml
@@ -41,6 +41,11 @@
             <groupId>com.alibaba</groupId>
             <artifactId>druid</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.liquibase</groupId>
+            <artifactId>liquibase-core</artifactId>
+            <optional>true</optional>
+        </dependency>
 
         <!-- Supported connected databases -->
         <dependency>
